social-app-backend/requirements.txt
fastapi[standard]==0.115.3
uvicorn==0.31.0
sqlalchemy==2.0.35
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
redis==5.1.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
colorlog==6.8.2
asyncpg==0.29.0
cloudinary==1.41.0
python-multipart==0.0.12
pydantic-settings==2.6.1
alembic==1.14.0
firebase-admin==6.6.0
celery==5.4.0
fastapi-mail